+++ src/controllers/genres.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Genres from "../models/genres";
 import { castSchema, updateCastSchema } from "../schemas/castSchema";
 import { genresSchema, updateGenresSchema } from "../schemas/genresSchema";
@@ -19,6 +20,12 @@ export const get = async (req, res) => {
 export const getById = async (req, res) => {
   try {
     const id = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({
+        message: "Invalid genres ID",
+      });
+    }
+
     const genres = await Genres.findById(id);
     if (!genres) {
       return res.status(404).json({
@@ -39,17 +46,21 @@ export const getById = async (req, res) => {
 export const remove = async (req, res) => {
   try {
     const id = req.params.id;
-    const genres = await Genres.findByIdAndRemove(id);
-    if (genres) {
-      res.status(204).send({
-        message: "Delete successfully",
-        data: genres,
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({
+        message: "Invalid movie ID",
       });
-    } else {
-      res.status(404).send({
+    }
+    const genres = await Genres.findByIdAndRemove(id);
+    if (!genres) {
+      return res.status(404).send({
         message: "The genres does not exist",
       });
     }
+    return res.status(204).send({
+      message: "Delete successfully",
+      data: genres,
+    });
   } catch (error) {
     res.status(500).send({
       message: error,
@@ -60,25 +71,28 @@ export const remove = async (req, res) => {
 export const update = async (req, res) => {
   try {
     const id = req.params.id;
+    if (!mongoose.Types.isValid(id)) {
+      return res.status(400).json({
+        message: "Invalid genres ID",
+      });
+    }
     const body = req.body;
     const { error } = updateGenresSchema.validate(body);
     if (error) {
-      res.status(400).send({
+      return res.status(400).send({
         message: error.message,
       });
-    } else {
-      const data = await Genres.findByIdAndUpdate(id, body, { new: true });
-      if (data) {
-        res.send({
-          message: "Update successfully",
-          data,
-        });
-      } else {
-        res.status(400).send({
-          message: "Genres is not existed",
-        });
-      }
     }
+    const data = await Genres.findByIdAndUpdate(id, body, { new: true });
+    if (!data) {
+      return res.status(400).send({
+        message: "Genres is not existed",
+      });
+    }
+    return res.send({
+      message: "Update successfully",
+      data,
+    });
   } catch (error) {
     res.status(500).send({
       message: err,
@@ -92,16 +106,15 @@ export const create = async (req, res) => {
 
     const { error } = genresSchema.validate(body);
     if (error) {
-      res.status(400).send({
+      return res.status(400).send({
         message: error.message,
       });
-    } else {
-      const data = await Genres.create(body);
-      res.send({
-        message: "Create cast successfully",
-        data,
-      });
     }
+    const data = await Genres.create(body);
+    return res.send({
+      message: "Create cast successfully",
+      data,
+    });
   } catch (err) {
     res.status(500).send({
       message: "Loi server",
