+++ src/controllers/movie.js
@@ -1,10 +1,23 @@
 import _ from "lodash";
 import Movie from "../models/movie";
 import { movieSchema, updateMovieSchema } from "../schemas/movieSchema";
+import mongoose from "mongoose";
 
 export const get = async (req, res) => {
   try {
-    const data = await Movie.find();
+    const data = await Movie.find().populate([
+      {
+        path: "casts",
+        select: ["name", "birthday"],
+      },
+      {
+        path: "genres",
+        select: "name",
+      },
+    ]);
+
+    // const movies = await Movie.find().populate("casts", "name");
+    // const data = await Movie.findOne({ title: "Phim Doraemon" }, { cast: 1 });
     res.send({
       message: "Get movies successfully",
       data,
@@ -16,22 +29,73 @@ export const get = async (req, res) => {
   }
 };
 
+// export const getById = async (req, res) => {
+//   try {
+//     const id = req.params.id;
+//     const movie = await Movie.findById(id).populate([
+//       {
+//         path: "casts",
+//         select: ["name", "birthday"],
+//       },
+//       {
+//         path: "genres",
+//         select: "name",
+//       },
+//     ]);
+
+//     console.log(movie, "movie ne");
+
+//     if (!movie) {
+//       return res.status(404).json({
+//         message: "The movie does not exist",
+//       });
+//     }
+//     return res.status(200).send({
+//       message: "Get movie successfully",
+//       data: movie,
+//     });
+//   } catch (error) {
+//     res.status(500).send({
+//       message: error,
+//     });
+//   }
+// };
+
 export const getById = async (req, res) => {
   try {
     const id = req.params.id;
-    const movie = await Movie.findById(id);
+
+    // Kiểm tra nếu id không hợp lệ
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({
+        message: "Invalid movie ID",
+      });
+    }
+
+    const movie = await Movie.findById(id).populate([
+      {
+        path: "casts",
+        select: ["name", "birthday"],
+      },
+      {
+        path: "genres",
+        select: "name",
+      },
+    ]);
+
     if (!movie) {
       return res.status(404).json({
         message: "The movie does not exist",
       });
     }
+
     return res.status(200).send({
       message: "Get movie successfully",
       data: movie,
     });
   } catch (error) {
     res.status(500).send({
-      message: error,
+      message: error.toString(),
     });
   }
 };
@@ -39,20 +103,27 @@ export const getById = async (req, res) => {
 export const remove = async (req, res) => {
   try {
     const id = req.params.id;
-    const movie = await Movie.findByIdAndRemove(id);
-    if (movie) {
-      res.status(204).send({
-        message: "Delete successfully",
-        data: movie,
+    // Kiểm tra nếu id không hợp lệ
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({
+        message: "Invalid movie ID",
       });
-    } else {
-      res.status(404).send({
+    }
+
+    const movie = await Movie.findByIdAndRemove(id);
+
+    if (!movie) {
+      return res.status(404).send({
         message: "The movie does not exist",
       });
     }
+    return res.status(204).send({
+      message: "Delete successfully",
+      data: movie,
+    });
   } catch (error) {
     res.status(500).send({
-      message: error,
+      message: error.toString(),
     });
   }
 };
@@ -61,20 +132,20 @@ export const create = async (req, res) => {
   try {
     const body = req.body;
     const { error } = movieSchema.validate(body);
+
     if (error) {
-      res.status(400).send({
+      return res.status(400).send({
         message: error.message,
       });
-    } else {
-      const data = await Movie.create(body);
-      res.send({
-        message: "Create movie successfully",
-        data,
-      });
     }
+    const data = await Movie.create(body);
+    return res.send({
+      message: "Create movie successfully",
+      data,
+    });
   } catch (err) {
     res.status(500).send({
-      message: "Loi server",
+      message: err.toString(),
     });
   }
 };
@@ -82,25 +153,31 @@ export const create = async (req, res) => {
 export const update = async (req, res) => {
   try {
     const id = req.params.id;
+    // Kiểm tra nếu id không hợp lệ
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({
+        message: "Invalid movie ID",
+      });
+    }
     const body = req.body;
+
     const { error } = updateMovieSchema.validate(body);
+
     if (error) {
-      res.status(400).send({
+      return res.status(400).send({
         message: error.message,
       });
-    } else {
-      const data = await Movie.findByIdAndUpdate(id, body, { new: true });
-      if (data) {
-        res.send({
-          message: "Update successfully",
-          data,
-        });
-      } else {
-        res.status(400).send({
-          message: "Movie is not existed",
-        });
-      }
     }
+    const data = await Movie.findByIdAndUpdate(id, body, { new: true });
+    if (!data) {
+      return res.status(400).send({
+        message: "Movie is not existed",
+      });
+    }
+    return res.send({
+      message: "Update successfully",
+      data,
+    });
   } catch (error) {
     res.status(500).send({
       message: error,
@@ -123,9 +200,18 @@ export const search = async (req, res) => {
     const movies = await Movie.find({
       $or: [
         { title: { $regex: query, $options: "i" } },
-        // { description: { $regex: query, $options: "i" } },
+        // { extract: { $regex: query, $options: "i" } },
       ],
-    });
+    }).populate([
+      {
+        path: "casts",
+        select: ["name", "birthday"],
+      },
+      {
+        path: "genres",
+        select: "name",
+      },
+    ]);
 
     res.status(200).json({
       message: "Search movies successfully",
@@ -138,27 +224,25 @@ export const search = async (req, res) => {
   }
 };
 
-const compareDates = (a, b) => {
-  const dateA = new Date(a.createdAt);
-  const dateB = new Date(b.createdAt);
-
-  if (dateA < dateB) {
-    return -1;
-  } else if (dateA > dateB) {
-    return 1;
-  } else {
-    return 0;
-  }
-};
-
 export const getSortedByCreatedAt = async (req, res) => {
   try {
-    const sortOrder = req.query.sortOrder || "desc"; // Mặc định là sắp xếp mới nhất
+    const year = req.query.year;
 
-    const movies = await Movie.find().sort({ createdAt: sortOrder });
+    const movies = await Movie.find({
+      year: { $gte: new Date(+year) },
+    }).populate([
+      {
+        path: "casts",
+        select: ["name", "birthday"],
+      },
+      {
+        path: "genres",
+        select: "name",
+      },
+    ]);
 
     res.status(200).json({
-      message: `Get movies sorted by created time (${sortOrder}) successfully`,
+      message: `Lấy các phim được tạo từ năm ${year} thành công`,
       data: movies,
     });
   } catch (error) {
